Migrate to C++ and fix preset serialization
- Rename all .c files to .cpp, remove extern "C" wrappers
- Fix preset save bug: only serialize waveformCount waveforms instead
  of all 8 slots (was writing uninitialized memory as garbage JSON)
- Add default member initializers to WaveformConfig struct
- Use NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT for forward-
  compatible preset loading (old presets load when new fields added)
- Remove redundant WaveformConfigDefault() function
- Suppress warnings from external headers (raylib, raygui)
- Fix narrowing conversion warning in ui_layout.cpp

+++ src/ui_layout.cpp
@@ -54,7 +54,7 @@ void UILayoutGroupBegin(UILayout* l, const char* title)
 void UILayoutGroupEnd(UILayout* l)
 {
     int groupH = (l->y + l->rowHeight + l->padding) - l->groupStartY;
-    GuiGroupBox((Rectangle){l->x, l->groupStartY, l->width, groupH}, l->groupTitle);
+    GuiGroupBox((Rectangle){(float)l->x, (float)l->groupStartY, (float)l->width, (float)groupH}, l->groupTitle);
     l->y = l->groupStartY + groupH + l->spacing * 2;
     l->rowHeight = 0;
 }
